PM-skill/requirements.txt
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.28.0
flask>=3.0.0
requests>=2.31.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
